converting int to pointer is 64bits risky, better use long even for dummy
git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3704 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 8f3410961103632c857d3abc24dc27f624294d9f

+++ platforms/iOS/plugins/SqueakFFIPrims/dummyFFI.c
@@ -157,7 +157,7 @@ int ffiAlloc(int byteSize)
 
 /*  ffiFree:
 Free space from the external heap */
-int ffiFree(int pointer)
+int ffiFree(long pointer)
 {
 return 1;
 }
